1. Swagger 적용 2. Exception, BindingResult 공통처리 3. build.gradle 정리 4. application.yml local, dev, prod 별로 구분 5. Entity에 radio, check Enum 적용 6. 게시판 첨부파일 업로드, 다운로드 적용

+++ src/main/java/reservation/pms/domain/lyhBoard/BoardRepository.java
@@ -7,7 +7,7 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import java.util.List;
 
-public interface BoardRepository extends JpaRepository<Board, Integer>, BoardRepositoryCustom {
+public interface BoardRepository extends JpaRepository<Board, Long>, BoardRepositoryCustom {
 
 //    리턴타입 {접두어}{도입부}By{프로퍼티 표현식}(조건식)[(And|Or){프로퍼티표현식}(조건식)]{정렬조건}(매개변수)
 //    리턴 타입 : E, Optional<E>, List<E>, Page<E>, Slice<E>, Stream<E>
